% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunes.R
\name{toLatex.stoicfortune}
\alias{toLatex.stoicfortune}
\title{Convert fortunes to Latex}
\usage{
\method{toLatex}{stoicfortune}(object, number = FALSE, width = c(1, 0.85),
  ...)
}
\arguments{
\item{object}{Object to convert}

\item{number}{Logical. Do you want the fortune number printed?}

\item{width}{integer specifying the character width.  By default
\code{getOption("width")} is used.}

\item{...}{Additional arguments.  Does nothing.}
}
\description{
Convert fortunes to Latex
}

